Fix the colors and static importing
Alright, so python is weird. If something is defined in __init__.py, it
will be run when the module itself is imported. That's fine. However, it
will also run when I import one if it's sub-modules, that is also run. I
have removed the imports from the static __init__.py file, so that we
can safely import `hammercloud.static.const` without running into issues
in the other static things we don't care about.